+++ ostwallet/src/main/java/com/ost/ostwallet/ui/dashboard/DashboardActivity.java
@@ -141,6 +141,7 @@ public class DashboardActivity extends BaseActivity implements
                     this);
             mViewPager.setCurrentItem(1);
         } else if(ostUser.getCurrentDevice().canBeAuthorized()) {
+            handleCrashAnalytics();
             mSettingsFragment.setOpenDeviceAuthorization(true);
             mViewPager.setCurrentItem(2);
         } else {
@@ -151,6 +152,7 @@ public class DashboardActivity extends BaseActivity implements
                     )) {
                 notifyActivate();
             }
+            handleCrashAnalytics();
         }
     }
 
@@ -198,6 +200,7 @@ public class DashboardActivity extends BaseActivity implements
     @Override
     public void activateAcknowledged(long workflowId) {
         FragmentUtils.goBack(this);
+        handleCrashAnalytics();
     }
 
     @Override
@@ -437,5 +440,68 @@ public class DashboardActivity extends BaseActivity implements
                         }
                     }
                 }, timeInMilliseconds);
+    }
+
+    private void handleCrashAnalytics() {
+        final AppProvider.FabricStateProvider fabricStateProvider = AppProvider.get().getFabricStateProvider();
+        fabricStateProvider.getUserDeviceFabricSetting(new AppProvider.FabricStateProvider.Callback() {
+            @Override
+            public void returnedPreference(Integer preference) {
+                if (preference == -1) {
+                    android.app.AlertDialog.Builder builder = new android.app.AlertDialog.Builder(AppProvider.get().getCurrentActivity());
+                    builder.setTitle("Crash Reporting");
+                    builder.setMessage("Would you like to share crash reports with OST to help improve the app?");
+                    builder.setCancelable(false);
+
+                    builder.setPositiveButton("Opt in", new DialogInterface.OnClickListener() {
+                        public void onClick(DialogInterface dialog, int id) {
+                            fabricStateProvider.setUserDeviceFabricSetting(true, new AppProvider.FabricStateProvider.Callback() {
+                                @Override
+                                public void returnedPreference(Integer preference) {
+                                    if (!fabricStateProvider.isFabricOn()) {
+                                        Fabric.with(AppProvider.get().getApplicationContext(), new Crashlytics());
+                                        fabricStateProvider.setFabricOn(true);
+                                        mSettingsFragment.reDrawView();
+                                    }
+                                }
+                            });
+                        }
+                    });
+                    builder.setNegativeButton("Opt out", new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+
+                            fabricStateProvider.setUserDeviceFabricSetting(false, new AppProvider.FabricStateProvider.Callback() {
+                                @Override
+                                public void returnedPreference(Integer preference) {
+                                    if (fabricStateProvider.isFabricOn()) {
+                                        fabricStateProvider.setFabricOn(false);
+                                        mSettingsFragment.reDrawView();
+                                        String title = "Opt out from crash reporting";
+                                        DialogFactory.createSimpleOkErrorDialog(AppProvider.get().getCurrentActivity(), title,
+                                                "For the changes to take effect, please exit the app and re-launch it").show();
+                                    }
+                                }
+                            });
+                        }
+                    });
+                    builder.create().show();
+                } else if (preference == 0) {
+                    if (fabricStateProvider.isFabricOn()) {
+                        fabricStateProvider.setFabricOn(false);
+                        mSettingsFragment.reDrawView();
+                        String title = "Opt out from crash reporting";
+                        DialogFactory.createSimpleOkErrorDialog(AppProvider.get().getCurrentActivity(), title,
+                                "For the changes to take effect, please exit the app and re-launch it").show();
+                    }
+                } else if (preference == 1){
+                    if (!fabricStateProvider.isFabricOn()) {
+                        Fabric.with(AppProvider.get().getApplicationContext(), new Crashlytics());
+                        fabricStateProvider.setFabricOn(true);
+                        mSettingsFragment.reDrawView();
+                    }
+                }
+            }
+        });
     }
 }
