+++ ostwallet/src/main/java/com/ost/ostwallet/ui/dashboard/WalletPresenter.java
@@ -164,67 +164,4 @@ class WalletPresenter extends BasePresenter<WalletView> implements
     public TransactionRecyclerViewAdapter getTransactionRecyclerViewAdapter() {
         return mTransactionRecyclerViewAdapter;
     }
-
-    public void onResume() {
-        handleCrashAnalytics();
-    }
-
-    private void handleCrashAnalytics() {
-        final AppProvider.FabricStateProvider fabricStateProvider = AppProvider.get().getFabricStateProvider();
-        fabricStateProvider.getUserDeviceFabricSetting(new AppProvider.FabricStateProvider.Callback() {
-            @Override
-            public void returnedPreference(Integer preference) {
-                if (preference == -1) {
-                    android.app.AlertDialog.Builder builder = new android.app.AlertDialog.Builder(AppProvider.get().getCurrentActivity());
-                    builder.setTitle("Crash Reporting");
-                    builder.setMessage("Would you like to share crash reports with OST to help improve the app?");
-
-                    builder.setPositiveButton("Opt in", new DialogInterface.OnClickListener() {
-                        public void onClick(DialogInterface dialog, int id) {
-                            fabricStateProvider.setUserDeviceFabricSetting(true, new AppProvider.FabricStateProvider.Callback() {
-                                @Override
-                                public void returnedPreference(Integer preference) {
-                                    if (!fabricStateProvider.isFabricOn()) {
-                                        Fabric.with(AppProvider.get().getApplicationContext(), new Crashlytics());
-                                        fabricStateProvider.setFabricOn(true);
-                                    }
-                                }
-                            });
-                        }
-                    });
-                    builder.setNegativeButton("Opt out", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-
-                            fabricStateProvider.setUserDeviceFabricSetting(false, new AppProvider.FabricStateProvider.Callback() {
-                                @Override
-                                public void returnedPreference(Integer preference) {
-                                    if (fabricStateProvider.isFabricOn()) {
-                                        fabricStateProvider.setFabricOn(false);
-                                        String title = "Opt out from crash reporting";
-                                        DialogFactory.createSimpleOkErrorDialog(AppProvider.get().getCurrentActivity(), title,
-                                                "For the changes to take effect, please exit the app and re-launch it").show();
-                                    }
-                                }
-                            });
-                        }
-                    });
-                    builder.create().show();
-                } else if (preference == 0) {
-                    if (fabricStateProvider.isFabricOn()) {
-                        fabricStateProvider.setFabricOn(false);
-                        String title = "Opt out from crash reporting";
-                        DialogFactory.createSimpleOkErrorDialog(AppProvider.get().getCurrentActivity(), title,
-                                "For the changes to take effect, please exit the app and re-launch it").show();
-                    }
-                } else if (preference == 1){
-                    if (!fabricStateProvider.isFabricOn()) {
-                        Fabric.with(AppProvider.get().getApplicationContext(), new Crashlytics());
-                        fabricStateProvider.setFabricOn(true);
-                    }
-                }
-            }
-        });
-    }
-
 }
